+++ ThirdPartyAdapters/vungle/vungle/src/main/java/com/vungle/mediation/VungleAdapter.java
@@ -2,10 +2,9 @@ package com.vungle.mediation;
 
 import androidx.annotation.Keep;
 import com.google.ads.mediation.vungle.VungleMediationAdapter;
-import com.google.android.gms.ads.mediation.MediationAdapter;
 
 /**
- * A {@link MediationAdapter} to load and show Vungle rewarded video ads using Google Mobile Ads SDK
+ * Adapter to load and show Vungle rewarded video | Native and RTB ads using Google Mobile Ads SDK
  * mediation.
  */
 @Keep