+++ ThirdPartyAdapters/vungle/CHANGELOG.md
@@ -4,7 +4,7 @@
 - new design.
 
 Built and tested with:
-- Google Mobile Ads SDK version 21.5.0.
+- Google Mobile Ads SDK version 22.0.0.
 - Vungle SDK version 7.0.0.
 
 #### Version 6.12.1.0